+++ packages/datasource/src/index.ts
@@ -1,3 +1,1 @@
-export * from "./core/index";
-export * from "./application/index";
-export * from "./utils/index";
+export * from "./core";