+++ packages/datasource/src/postgres/index.ts
@@ -1,0 +1,11 @@
+export type {
+  DbConfig,
+  PostgresOrgNodeRow,
+  PostgresOrgScopeData,
+  PostgresOrgScopeInput,
+  PostgresRoleDataPolicyRow
+} from "../core/interfaces";
+export * from "./config";
+export * from "./postgres-datasource.factory";
+export * from "./postgres.adapter";
+export * from "./postgres-org-scope.adapter";